Revisit the architecture to support a hierarchical menu.
This commit implements the hierarchical menu & remote. The views
are refactored to support the recursive tree definition. The
opportunity is taken to remove sinatra/mustache and directly use
View classes as pure ruby objects and no sinatra magic.

The menu behavior (next, previous, jump, skip unscened nodes) is
implemented through a Walk class helper. The latter captures a
linearization of the nodes as well as a current state, and provides
an immutable API to walk the tree with next, previous and jump
operations. The main controller keeps its state as a Walk instance.

The Walk instance, especially `jump`  relies on nodes being uniquely
identified according to a DFS. This is ensured by the Config model
class itself, that identifies nodes according to the same linearization
strategy.