Improve the menu and remote to avoid loosing vertical context.
This commit improves both the menu and the remote for situations
where the number of pages is such that the menu does not fit vertically
on the screen. In such cases, the refreshing mechanism was such that
the user ended up on top of the menu on every page change.

The commit improves this by adding a refresh callback to the refresh
helper function. The callback is responsible for actually refreshing
the page when it changes. Two different strategies are used for the
display and remote pages. The remote hacks on CSS classes without
full reloading the page. The display still does it.

The menu is also improved to highlight the current page on opening.