Improve cache control of assets & pages.
This commits splits the controller and the assets/webassets service
into two different sinatra applications, without any cache contro.
A main Service instance is build from both of them and adds
a shared Rack::Nocache that controls the cache headers.

The caching is such that assets served with either a Last-Modified
or an ETag header allow public cache for one hour, with a
revalidation required after that.