Add support for play/pause from the remote control.
This commit adds a rufus scheduler to the architecture, to loop over
scenes through an automated walk. The animation frequency and behavior
at start can be configured in the main scenes.yml file. The animation
can also be paused and resumed from the remote control.

The implementation refactors the notion of 'state' to not only include
the walk index, but also the scheduler state (paused or playing). The
refresh.js is changed accordingly. This was necessary to display the
correct icon (play vs. pause) in the remote.