Let a node be repeated x times in auto play mode. Refs #27, #29.
This is a way to specify the duration of each node, using a
`repeat` entry in its configuration.

This commit adds an optional field 'repeat' to the node
section of the easycast configuration file schema. It is an
integer greater than 1 that specifies the number of time
that a node shall be repeated when it is casted automatically
(play is on). It was added to let the user able to specify
distinct durations for each node.

Technically the commit distinguishes between an external
Walk state that is the so-called state of the walk, i.e.
the node of the tree that the walk is currently displaying,
and an internal state that is the node and the remaining
number of time that the node shall be displayed through
the 'next' method. When the repeat parameter is 1 or less,
the next method acts normally (goes to next node). Previous,
jump and next with auto = false all reinitialize to the number
specified in the config file for this node.