Implement the gallery with only one image tag.
This commit refactors the gallery to use another strategy. The
reason is that the Raspberry PI has a hard time playing existing
animations and usually uses 100% CPU during it. This is most
probably due to chromium trying to stack all images with
transparency before starting the animation.

This commit changes the strategy to use a single img tag that
displays the current image. It should make the browser work much
easier.